backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
httplib2==0.22.0
httptools==0.6.4
huggingface-hub==0.34.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
openai==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.2.3
pytesseract==0.3.13
pytest==8.4.2
pytest-asyncio==1.1.0
python-dotenv==1.1.1
python-magic-bin==0.4.14
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.7.34
reportlab==4.4.3
rsa==4.9.1
safetensors==0.6.1
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.55.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
slowapi==0.1.9
limits==5.6.0
httpx==0.27.0
python-jose==3.3.0
celery==5.4.0
redis==5.0.8
prometheus-fastapi-instrumentator==7.0.0
SQLAlchemy==2.0.36